move redis config to application.java
