+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/config/RedisConfig.java
@@ -11,29 +11,4 @@ import org.springframework.data.redis.core.RedisTemplate;
 @Configuration
 public class RedisConfig {
 
-  @Bean
-  protected JedisConnectionFactory jedisConnectionFactory() {
-    RedisStandaloneConfiguration configuration =
-        new RedisStandaloneConfiguration("localhost", 6379);
-    configuration.setDatabase(3);
-    JedisClientConfiguration jedisClientConfiguration =
-        JedisClientConfiguration.builder().usePooling().build();
-    JedisConnectionFactory factory =
-        new JedisConnectionFactory(configuration, jedisClientConfiguration);
-    factory.afterPropertiesSet();
-    return factory;
-  }
-
-  @Bean
-  public RedisTemplate<String, Object> redisTemplate(
-      JedisConnectionFactory jedisConnectionFactory) {
-    final RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-    redisTemplate.setConnectionFactory(jedisConnectionFactory);
-    return redisTemplate;
-  }
-
-  @Bean
-  public RedisConnectionFactory redisConnectionFactory() {
-    return jedisConnectionFactory();
-  }
 }
